+++ media/main.js
@@ -45,6 +45,7 @@ $(document).ready(function () {
       )
       .replaceAll("/@", "/");
   };
+
   const addPRsToDoc = (prs) => {
     $("#ghHolder").append(
       "<button class='run-watermelon'>Run Watermelon</button><br/>"
@@ -76,7 +77,7 @@ $(document).ready(function () {
           </h5>
         </div>
         <div class="comment-body">
-      ${replaceUserTags(marked.parse(comment.body))}
+      ${comment?.body ? replaceUserTags(marked.parse(comment.body)) : ""}
         </div>
         </div>`;
       });
@@ -90,7 +91,7 @@ $(document).ready(function () {
         <div>
           <div class="pr-owner">
             <p class="pr-poster" title="View this user on github">
-              Author: <a href="${pr.userLink}">${pr.user}</a>
+              <a class="pr-author-combo" href="${pr.userLink}"><img class='pr-author-img' src="${pr.userImage}" />${pr.user}</a>
             </p>
             <p class="pr-date">
               ${new Date(pr.created_at).toLocaleDateString("en-us", {
@@ -102,10 +103,12 @@ $(document).ready(function () {
             </p>
           </div>
           <div class="pr-body">
-            ${replaceIssueLinks(
+            ${pr?.body ?
+              replaceIssueLinks(
               replaceUserTags(marked.parse(pr.body)),
-              pr.repo_url
-            )}
+              pr.repo_url)
+              : ""
+            }
           </div>
           ${mdComments}
         </div>