+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Fixed some instances of the get help on Slack button not working
 - Will select whole file if no selection is made
 - Added a wildcard allow to watermelon.tools in CSP
+- Allow all of github as source for images
+- Fixed case where empty PR body or empty comment would break the app
+- Fixed bug that retrieved PRs from unrelated repositories
+- Added Connect-src to CSP
+- Added PR author image
 
 ## [1.1.3]
 - Fixed some instances of the "Get help on Slack" running search